add missing part of the variable type changes in do_gotoline()
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1854 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/search.c
@@ -846,7 +846,7 @@ void do_gotoline(ssize_t line, int save_pos)
 	    return;
 	}
 
-	line = atoi(answer);
+	line = (ssize_t)atol(answer);
 
 	/* Bounds check. */
 	if (line <= 0) {
